+++ frontend/plugin/src/lib/epistola.module.ts
@@ -16,6 +16,7 @@ import {
 import {DataMappingTreeComponent} from './components/data-mapping-tree/data-mapping-tree.component';
 import {FieldTreeComponent} from './components/field-tree/field-tree.component';
 import {EpistolaDownloadComponent} from './components/epistola-download/epistola-download.component';
+import {EpistolaRetryFormComponent} from './components/epistola-retry-form/epistola-retry-form.component';
 import {EpistolaPluginService} from './services';
 
 @NgModule({
@@ -32,7 +33,8 @@ import {EpistolaPluginService} from './services';
     DownloadDocumentConfigurationComponent,
     DataMappingTreeComponent,
     FieldTreeComponent,
-    EpistolaDownloadComponent
+    EpistolaDownloadComponent,
+    EpistolaRetryFormComponent
   ],
   exports: [
     EpistolaConfigurationComponent,
@@ -41,7 +43,8 @@ import {EpistolaPluginService} from './services';
     DownloadDocumentConfigurationComponent,
     DataMappingTreeComponent,
     FieldTreeComponent,
-    EpistolaDownloadComponent
+    EpistolaDownloadComponent,
+    EpistolaRetryFormComponent
   ],
   providers: [
     EpistolaPluginService